# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  category: string;
  description: string;
  date: string;
  paymentMethod: string;
  notes?: string;
  createdAt?: string;
}

export type CreateTransactionType =
  (typeof CreateTransactionType)[keyof typeof CreateTransactionType];

export const CreateTransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransaction {
  type: CreateTransactionType;
  amount: number;
  category: string;
  description: string;
  date: string;
  paymentMethod: string;
  notes?: string;
}

export interface Budget {
  id: number;
  category: string;
  limit: number;
  spent: number;
  month: string;
}

export interface CreateBudget {
  category: string;
  limit: number;
  month: string;
}

export type FinancialSummaryCategoryBreakdownItem = {
  category: string;
  amount: number;
};

export type FinancialSummaryMonthlyTrendItem = {
  month: string;
  income: number;
  expenses: number;
};

export interface FinancialSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  categoryBreakdown: FinancialSummaryCategoryBreakdownItem[];
  monthlyTrend: FinancialSummaryMonthlyTrendItem[];
}

export type ListTransactionsParams = {
  type?: ListTransactionsType;
  category?: string;
  search?: string;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
  all: "all",
} as const;