# lib/api-zod/src/generated/types/createTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionType } from "./createTransactionType";

export interface CreateTransaction {
  type: CreateTransactionType;
  amount: number;
  category: string;
  description: string;
  date: string;
  paymentMethod: string;
  notes?: string;
}