# lib/api-zod/src/generated/types/financialSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialSummaryCategoryBreakdownItem } from "./financialSummaryCategoryBreakdownItem";
import type { FinancialSummaryMonthlyTrendItem } from "./financialSummaryMonthlyTrendItem";

export interface FinancialSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  categoryBreakdown: FinancialSummaryCategoryBreakdownItem[];
  monthlyTrend: FinancialSummaryMonthlyTrendItem[];
}