# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  category: string;
  description: string;
  date: string;
  paymentMethod: string;
  notes?: string;
  createdAt?: string;
}